% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxx.create_and_setwd.R
\name{fxx.create_and_setwd}
\alias{fxx.create_and_setwd}
\title{Create directory (if needed) and set working directory}
\usage{
fxx.create_and_setwd(path)
}
\arguments{
\item{path}{Target directory path.}
}
\value{
Invisibly returns normalized path.
}
\description{
Create directory (if needed) and set working directory
}
