% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxx.get_many_color.R
\name{fxx.get_many_color}
\alias{fxx.get_many_color}
\title{Get many colors from ggsci palettes}
\usage{
fxx.get_many_color(num = 10)
}
\arguments{
\item{num}{Number of colors requested.}
}
\value{
A character vector of hex colors.
}
\description{
Get many colors from ggsci palettes
}
