% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxx.sanitize_variable_name.R
\name{fxx.sanitize_variable_name}
\alias{fxx.sanitize_variable_name}
\title{Sanitize text into a valid-ish variable name}
\usage{
fxx.sanitize_variable_name(input_string)
}
\arguments{
\item{input_string}{Character vector.}
}
\value{
Character vector with non-valid characters replaced by "_".
}
\description{
Sanitize text into a valid-ish variable name
}
