% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fxx.baseplot_obj2.R
\name{fxx.save_plot}
\alias{fxx.save_plot}
\title{Save a plot object to PNG and PDF}
\usage{
fxx.save_plot(plot, filename, height, width, dpi = 600)
}
\arguments{
\item{plot}{A plot object. Supports ggplot, grid grob/gList, pheatmap result,
and ComplexHeatmap objects.}

\item{filename}{Output file path without extension.}

\item{height, height_plot}{Plot height in inches.}

\item{width, width_plot}{Plot width in inches.}

\item{dpi}{Resolution for PNG.}
}
\value{
Invisibly returns NULL.
}
\description{
Save a plot object to PNG and PDF
}
